Well on the way to completely messing up the STEP code, just by flipping the Y coord!

+++ src/hid/gtk/edge3d.h
@@ -5,13 +5,13 @@ typedef struct
 
   /* For circular curves */
   bool is_round;
-  float cx;
-  float cy;
-  float cz;
-  float nx;
-  float ny;
-  float nz;
-  float radius;
+  double cx;
+  double cy;
+  double cz;
+  double nx;
+  double ny;
+  double nz;
+  double radius;
 
   /* STEP crap - to hell with encapsulation */
   step_id infinite_line_identifier;
@@ -19,6 +19,6 @@ typedef struct
 } edge_info;
 
 edge_info *make_edge_info (void);
-void edge_info_set_round (edge_info *info, float cx, float cy, float cz, float nx, float ny, float nz, float radius);
+void edge_info_set_round (edge_info *info, double cx, double cy, double cz, double nx, double ny, double nz, double radius);
 void edge_info_set_stitch (edge_info *info);
 void destroy_edge_info (edge_info *info);
